+++ tests/Issues/DoNotReplaceUnknownClasses/config/configured_rule.php
